Increase double->string precision
When calling php library methods to convert doubles to strings
we were only keeping 8 digits after the decimal point.  For
double precision numbers, 15 are stable.

Redis uses the C library function strtod() to do the conversion,
which appears to use the first 16 digits for aproximation:
http://www.exploringbinary.com/how-strtod-works-and-sometimes-doesnt/

This hotfix increases the signifigant digits to 16, which I think
is the correct number to pick here.

Addresses issue #371