Initial commit of RedisCluster class and library
This is an initial commmit which adds a RedisCluster class as well
as the framework around which we'll be building proper cluster
support.

The first commit just contains the code to set up and use our new
RedisCluster class as well as parsing logic to handline CLUSTER NODES
such that we can map the keyspace.

Next up, command processing and then pipelining in a sane way.