Initial commit of redis_commands.c
Given that non cluster and cluster based redis objects should act
exactly the same way (in terms of argument handling/processing)
we're extracting the parsing of arguments and command construction
from being directly inside of the PHP_METHOD itself.

This way, we don't repeat ourselves having two identical places
where we do this but in a slightly different way.

Every command takes a RedisSock pointer (coming from redis_sock_get
in a standard redis class, and cluster->flags in a cluster).

If we're passed a pointer for slot, we'll hash one of the keys
to return it's proper slot.