Add lazy connect logic, use seeds not nodes for fallback.
We need to try to connect to our node if we've mapped it but not yet
connected to the server itself.

In addition, rather than iterating through nodes 0-CLUSTER_MAX_SLOT
as a fallback, just use our seeds, as in almost all cases many
slots will point to the same Redis instance.