Transaction support
This is the initial commit supporting transactions in RedisCluster
via MULTI..EXEC.  Given that the data can be distributed across
many nodes, we accomplish transactions as follows.

1.  When entering MULTI mode, the cluster context state is changed
    to MULTI
2.  When we send a command to a node, we check that node's RedisSock
    state, and if it's still ATOMIC, deliver the MULTI command,
    updating the node's state to MULTI.
3.  When reading replies, we check if the RedisSock* is in a MULTI
    state, which means we need to issue the EXEC command.  We do
    this and then flip the state back to ATOMIC.
4.  On completion, everything is reverted to ATOMIC state and our
    reply callbacks are freed and set to NULL.

For transactions, we enforce writes such that they MUST be able to
be processed by the node in question (e.g. they can't work during
a reshard, because there is no way to know what has failed, etc).

This is the same as RedisCluster which will also cancel transactions
in the case where a request was made for data that is somewhere else.