WATCH command
Implemented the WATCH command for RedisCluster.  This command can
take any number of keys, so phpredis splits the request across the
cluster in the best way it can.

For every key in a multiple key command, Redis Cluster requires
that they all hash to the same SLOT, else it will return a
CROSSLOT error and fail.

For WATCH in RedisCluster, a few things to note:

* The command will fail if phpredis is out of sync with the keyspace.
  This is because we'll need to know where to deliver each command,
  or can't possibly deliver them correctlhy.

* The command will fail if any command delivery failures occur on any
  node.  This is the case either for a normal communication error or
  if RedisCluster returns to us MOVED/ASK redirection.