Initial commit of MSET and MSETNX commands.
More work needs to be done to work through what, exactly, to do
in the case of a protocol error in commands like this.  If Redis
(or communication with Redis) bombs out halfway through a MSETNX
reply handler, we will have alredy sent the commands but not yet
consumed all of the responses.

This would put RedisCluster in an inconsistent state, and we have
a few options.

1.  Abort immediately
2.  Continue to process responses, but php_error_docref that it
    occured and use FALSE to indicate a failure occured.
3.  Stop processing responses, and reset the connection to Cluster

More testing will need to be done to figure out what the best
stragegy is.