Refactoring, updated handling of MULTI commands
Minor formatting changes

Updated how we delete context variables, such that our context
pointers aren't always going to be a zval** array.  This
introduces the possibility of creating a memory leak if we have
to abort a transaction due to a communications error, but this
can be handled later by attaching a void dtor(void*) callback.