Don't NULL out the original slot on MOVED error
There is no reason to NULL out the originally requested slot
when we get a MOVED error, because all we really know is that
the data (at slot N) that we're requesting is somewhere else.

Instead, just update the node mapping for the destination MOVED
slot, if it has changed.  The source mapping will change organically
as we send more requests.