EVAL/EVALSHA
Implemented EVAL and EVALSHA commands for cluster.  Thes commands
present a bit of an issue from the context of a client.  Given that
the code being executed can be any arbitrary LUA code, phpredis can't
always protect against trying to access keys that live on a different
cluster node, especially if the "keys count" isn't passed correctly,
or in cases where the keys are generated in the script.

Work in progress, but there is unlikely a perfect solution to this
at all, given the nature of Redis Cluster and LUA execution.