SCAN command
Initial commit of the SCAN command for cluster.  As a first iteration,
the command takes a host:port string (which can be retreived from the
introspection method RedisCluster::_masters()), to scan either
everything in the cluster (e.g. in a loop), or one in particular.

A simple way to scan everything in the whole cluster Keyspace

$obj_c->setOption(RedisCluster::OPT_SCAN, RedisCluster::SCAN_RETRY);
foreach($obj_c->_masters() as $str_hash) {
    $it = NULL;
    while($arr = $obj_c->scan($it, $str_hash)) {
        foreach($arr as $key) {
            echo "$str_hash -> " . $key . "\n";
        }
    }
}