Update how we call node directed commands
Initially when sending commands to a specific node (e.g. INFO,
BGSAVE, etc), the semantics were either to send a key or a host
and port in two arguments.  I think this would have been confusing
and caused issues with more complicated commands like PUBSUB or
SCRIPT.

Now, to send a command, one either passes a string key or an array
with host and port (which is what _masters now returns always).