Update redirection logic to throw on CLUSTERDOWN
When Redis Cluster responds with a CLUSTERDOWN error, we should
throw an exception.  It seems that in the current state of cluster
(during slave election), it will enter a CLUSTERDOWN state for a
short period of time.

This allows clients to trap the exception and then either wait and
try again, or do something else.