KEYS command
Keys should never be used in production, but this implementation
allows the use of KEYS in such a way that you can call it normally
and it will look through all of the known masters and return the
whole keyspace.

We might want to implement this only as an option (even at compile
time) as it's very dangerous to run on large keyspaces.