Support for the "command" command.
This is a simple addition that allows a client to call any given
Redis command by sending the command name followed by a list of
arguments.

This is useful for the cases where there are new commands in Redis
that have not yet been specifically implemented in phpredis, or
if you want to use phpredis as a pass-through where the commands
and arguments are coming from somewhere else (e.g. monitor logs, etc).