Added a new method sAddArray to both Redis and RedisCluster
Presently, the sAdd command is variadic, meaning that it takes a key
and then 1 to N additional arguments for the members being added.  We
need to keep this functionality to avoid breaking existing code,
but there are good performance and other reasons to have an sAdd
command which takes a key followed by an array of members, which is
what the sAddArray method implements.