Ensure we don't try and send 0 expires for SET
As raised by issues #579, and #640 phpredis could be made to silently
revert to a "SET" call (no expiry) if a non long argument was specified
for EX or PX options, as well as an invalid base timeout value.

This modification changes the code so we don't even send the command to
Redis in such cases, which also avoids a round trip.