Fixed many memory leaks.
The php7 api changed from previous versions in that there is no
longer the option to "duplicate or not duplicate" a string value
when using the ZVAL_STRING and ZVAL_STRINGL macros.

This means that every time ZVAL_STRING or ZVAL_STRINGL is being
called, one *must* call zval_dtor on that zval or there will be
a leak.

In addition there were several places where the return value from
a zend_call_function passthrough was not being freed, and then
being used again.  Specifically (in the context of RedisArray)
it was when the array was already in MULTI mode, meaning that the
return is a refcounted copy of the Redis object itself, which also
needs to be freed to avoid a leak.

All unit tests passing in debug/release mode with no memory leaks
being reported or any complaining from valgrind.