Initial commit for cluster slot caching.
This commit uses PHP's persistent_list to cache cluster slot
information within or between requests.  The benefit here is
that we no longer need to perform and parse CLUSTER SLOTS
each and every time a RedisCluster object is created.

This is still a work in progress, as the caching code should
probably be moved into it's own file, and much testing needs
to be done to make sure cache invalidation is handled properly.