Use static declarations for spl_ce_RuntimeException decl
This is consistent with the PDO extension, which also resolves the RuntimeException class entry on-demand. Without a static decl, this can conflict with the extern decls in spl_exceptions.h and cause a "missing symbol" error for other extensions depending on those SPL decls.

This commit also makes the spl_rte_ce decl in redis_cluster.c static for consistency, although that symbol does not currently conflict with anything in SPL.