Allow for string based longs in zrangebyscore
This works differently than the non-php7 version, as the zend library
has now added explicit methods for conversions such as this, which
handle more types (e.g. FALSE/NULL will equate to zero, and TRUE will
equate to 1).

However a non numeric string such as "not a number" will also equate
to zero, but this seems to be the way it's intended to work.

Fixes #906