Move zend_object handler to the end
In php7 the zend_object handler structure is inlined (is no longer a pointer,
but rather variable sized depending on various things, so it needs to be at the
end of the container class.

This is mentioned in the extensive /s upgrading documentation from Zend

See "custom objects":
https://wiki.php.net/phpng-upgrading

In addition I believe that the zend library now takes care of freeing the overall
structure, so that shouldn't be done anymore if running php >= 7.0.0.