Use zend_list_close instead of zend_list_delete when reconnecting.
In php7 zend_list_close handles closing the connection and frees
resource structures but doesn't decrease refcount (and therefore
remove the element from the list).

This segfault could be triggered in php7 by simply doing this:

```php
$r = new Redis();
$r->connect('localhost', 6379);
$r->connect('localhost', 6379);
$r->ping(); // or any other command
```

What appears to have been happening here, is that phpredis would
delete any existing socket resource on a connect call to an already
existing object, but then the call to add_property_resource would
trigger the destructor of the socket (as the Zend library already
saw a "socket" property).

This would set redis_sock->stream->abstract (the actual socket) to
NULL, causing any subsequent communication to segfault or issue a
bus error.

Fixes #1076