Add arginfo for ReflectionMethod to redis for some commands
This is incomplete, writing arginfo for everything in a single PR
would be difficult for reviewers.
This focuses on string operations and some common commands for
redis 1.0.

The macro's arguments are:
ZEND_BEGIN_ARG_INFO_EX(name, _unused, return_reference, required_num_args)

For #1055 (Still a lot of work)

This is based on what was written in README, not 100% sure if it matches
the code.