Treat a -1 response from cluster_check_response as a timeout.
When cluster_check_response returns -1 this can be treated as a
timeout.

Note that there is one non-timout condition which can
cause a -1 response, but that is a corrupted MOVE/ASK reply
which can *probably* be treated as a timeout as well,
because it means that something has gone horribly wrong
with the connection.

Addresses #1425