Fix XCLAIM for 32-bit PHP
This fix addresses #1437

There are two issues happening here.  First, the XCLAIM option TIME
expects a UNIXTIME in milliseconds, which cannot be represented in
a long on 32-bit PHP builds.  For that reason, I've created a helper
function to allow more flexibility to the user.

Now, you can pass us a PHP long, float, or string and we'll try to
convert the value to something Redis can use.  We make a best-effort to
inform the user if they have passed something invalid (e.g. negative
values, or something non-scalar).