Initial attempt at weighted session failover
This commit modifies our weighted session selection algorithm by adding
a scratch array so we can mark servers that we've picked but were unable
to connect to.

All we really do is treat any server that has failed as if it had a
weight of zero meaning it'll never be picked.

Addresses #1214