Ignore preferred nodes mode on MOVED response
Redis can be in a state where a replica will respond MOVED to a slot even if
that replica should be able to serve that slot. The why is immaterial - redis
spec says you must perform your command on the host:port the MOVED response
specifies.

If we always sort our preferred nodes higher than the host:port it says we
should read from - we end up never reading from it. This shortcuts that by
disabling the FAILOVER mode entirely when this situation is encounted -
ensuring the command then goes to the primary.

This has a slight benefit over FAILOVER_DISTRIBUTE_SLAVES mode - as that will
work, but randomly picks a node out of the list (primary+replicas) each time -
you just happen to eventually get the primary and have it continue. By
switching to NONE - we immediately just talk to the primary.