Test against valkey
Add Valkey to our server matrix in addition to making the jobs a bit
more efficient by only installing the specific server we're testing on
each run.

For now we allow tests to fail against Valkey as they don't yet have an
official release.  Once there is an official release we'll remove the
`continue-on-error` setting for Valkey.