Add a CI step that waits for server instances before running tests
Every so often our tests will fail because we attempt to interact with
one of the daemonized server instances before it has enough time to
actually start.

Usually this happens when we try to use the cli tool to execute
`--cluster-create`, but it can occur elsewhere as well.

This commit adds a step that waits for every instance that we started to
actually be up before trying to create the cluster and run subsequent
unit tests.

Additionally it switches from `$(seq a b)` to the `{a..b}` brace
expansion.