Rework how and when update slot cache and remap topology.
When a replica node is taken offline, it will briefly respond to
commands with some form of `CLUSTERDOWN` error. Presently while this did
cause PhpRedis to invalidate the slot cache, we did not immediately
remap the topology.

What this meant, is that for an already connected cluster, we might keep
sending readonly commands to this now decomissioned cluster replica for
the duration of the request.

Additionally, if the request was long lived enough we may still have it
mapped if the replica was brought back online at which point the client
might see dozens of `LOADING` exceptions.

This commit just changes our behavior when we detect any `CLUSTERDOWN`
error such that we flush the slot cache, disconnect the particular node,
and also attempt a remap of the topology.