Add `invalidateSlotCaches()` function to invalidate any slot caches.
This commit adds a mechanism to invalidate every `RedisCluster` slot
cache across all forked workers with one call to any of the workers.

The invalidation itslef works by using a `uint64_t` generation counter
which is allocated in shared memory on `MINIT` which is stored in every
slot cache when it is created.

When a user wants to invalidate any of the slot caches they can call
`RedisCluster::invalidateSlotCaches()` which simply does an atomic
increment on this global genertion value.

When a slot cache is retreived from the persistent list, the internal
generation is compared to the global shared memory generation and if
they don't match, the cache is not used, which is the same as
invalidting it as RedisCluster will write a new slot cache on object
destruction.

The feature is only compiled if `config.m4` determines that using `mmap`
to allocate shared memory is available and works.

Additionally, the feature must be explicitly enabled by setting the
`redis.clusters.shared_slot_cache_invalidation`.